+++ CHANGELOG.md
@@ -7,7 +7,12 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [Unreleased]
 
 ### Added
+
 - Extra git command args for getting logs for each line of each relevant file
+
+### Fixed
+
+- Don't try to render heatmap decorations when no commits were found
 
 ## [0.0.1] - 2020-10-07